refactor backend import graph into tree shape
