+++ backend/orchestrator.py
@@ -12,6 +12,7 @@ from types import MappingProxyType
 from typing import Any, AsyncIterator, List, Mapping, Optional, Sequence
 
 from . import database as db
+from .cached_call import CachedBase
 from .database.models import (
     ActiveLorebookEntryRow,
     CharacterCardRow,
@@ -23,9 +24,9 @@ from .database.models import (
     SettingsRow,
     UserPersonaRow,
 )
-from .kv_tracker import CachedBase, _KVCacheTracker
+from .kv_tracker import _KVCacheTracker
 from .llm_client import AbortToken, LLMClient, reasoning_cfg
-from .llm_types import ChatMessage, ContentPart
+from .llm_types import ChatMessage
 from .locks import workflow_character_state_lock, workflow_state_lock
 from .macros import Macros
 from .passes.director import (
@@ -42,6 +43,7 @@ from .passes.editor.length_guard import (
     resolve_length_guard,
 )
 from .passes.writer import writer_stage
+from .pipeline_state import ModelLane, TurnState, _PipelineConfig
 from .prompt_builder import (
     build_prefix,
     compute_lorebook_injection_block,
@@ -67,28 +69,6 @@ logger = logging.getLogger(__name__)
 # ── Core pipeline ─────────────────────────────────────────────────────────────
 
 
-@dataclass(frozen=True)
-class ModelLane:
-    """One model's call surface for the turn: a client paired with its
-    byte-identical cached bottom (prefix + tools + model + the macro ``resolve``
-    hook that scrubs placeholders from the final wire bytes).
-
-    A turn has two lanes — ``writer`` and ``agent`` (director + editor). In
-    single-model mode they are the *same object* (the writer's lane is reused for
-    the agent), so the byte-identity invariant "director + editor + writer ride
-    the same base" is structural, not a convention each call site must honour. In
-    dual-model mode they are distinct: the agent lane carries the agent server's
-    client, its own prefix + tool blob, and the agent model; the writer lane
-    carries the writer client with an empty tools blob (Invariant 5).
-
-    ``reasoning`` stays per-pass (director and editor share the agent lane but
-    toggle reasoning independently), so it is not part of the lane.
-    """
-
-    client: LLMClient
-    base: CachedBase
-
-
 def is_dual_model(agent_client: LLMClient | None) -> bool:
     """Return True when a separate agent endpoint is configured (dual-model mode).
 
@@ -107,26 +87,6 @@ def agent_enabled(settings: Mapping[str, Any]) -> bool:
     semantics stay consistent everywhere.
     """
     return bool(settings.get("enable_agent", 1))
-
-
-@dataclass
-class _PipelineConfig:
-    """Resolved per-turn flags, lanes, and prefixes for :func:`_run_pipeline`."""
-
-    agent_on: bool
-    enabled_tools: Mapping[str, bool]
-    director_reasoning_on: bool
-    writer_reasoning_on: bool
-    editor_reasoning_on: bool
-    audit_enabled: bool
-    length_guard: LengthGuard | None
-    do_edit: bool
-    writer_enabled_tools: Mapping[str, bool]
-    # The two call surfaces for the turn. ``writer_lane`` runs the writer pass;
-    # ``agent_lane`` runs director + editor. In single-model mode they are the
-    # same object by construction (see :class:`ModelLane`).
-    writer_lane: ModelLane
-    agent_lane: ModelLane
 
 
 def _resolve_pipeline_config(
@@ -235,51 +195,6 @@ class _PipelineResult:
         purpose: ``staged_attachments`` carries raw artifact bytes that must not
         be deep-copied."""
         return {f.name: getattr(self, f.name) for f in fields(self)}
-
-
-@dataclass
-class TurnState:
-    """Mutable per-turn state threaded by reference through the three pass
-    stages (``director_stage`` / ``writer_stage`` / ``editor_stage``).
-
-    These were ``_run_pipeline``'s ~20 turn-state locals. The result-bound
-    fields mirror :class:`_PipelineResult` (and :class:`DirectorResult`) names so
-    one name follows each value from the director pass through to persistence;
-    :func:`_make_result` reads this straight into a :class:`_PipelineResult`.
-
-    Seeded in :func:`_run_pipeline` from ``director`` (``active_moods`` and the
-    progressive seed filtered to valid fragment ids) and the resolved
-    ``user_message`` (``effective_msg``). ``progressive_state`` /
-    ``valid_progressive_ids`` are turn inputs (not result fields): the director
-    seed map and the id set used to filter director output into
-    ``progressive_fields``.
-    """
-
-    # --- seeds / inputs ---
-    user_message: str = ""
-    effective_msg: str = ""
-    active_moods: list[str] = field(default_factory=list)
-    progressive_state: dict = field(default_factory=dict)
-    valid_progressive_ids: set[str] = field(default_factory=set)
-
-    # --- director outputs ---
-    agent_raw: str = ""
-    calls: list[dict] = field(default_factory=list)
-    latency: int = 0
-    rewritten_msg: str | None = None
-    extra_fields: dict = field(default_factory=dict)
-    progressive_fields: dict = field(default_factory=dict)
-    selected_lorebook_entries: list[str] = field(default_factory=list)
-    inj_block: str = ""
-    writer_lorebook_block: str = ""
-
-    # --- writer / editor outputs ---
-    resp_text: str = ""
-    writer_content: "str | list[ContentPart]" = ""
-    reasoning_director: str = ""
-    reasoning_writer: str = ""
-    reasoning_editor: str = ""
-    feedback_values: dict = field(default_factory=dict)
 
 
 def _make_result(state: TurnState, staged: list[dict] | None = None, staged_state: dict | None = None) -> dict:
