+++ backend/passes/director/prompt_rewrite.py
@@ -12,9 +12,10 @@ overwrite — those I/O steps stay in the orchestrator; this module is pure.
 
 The tool *schema* deliberately stays in ``tool_registry.py`` (``REWRITE_PROMPT_TOOL``
 in the ``TOOLS`` registry): it is part of the cached tools blob sent to the LLM, so
-moving it would bust the KV cache. Only the instruction template and the Python
-glue around the tool relocate here — mirroring how ``length_guard.py`` leaves the
-``editor_rewrite`` schema in ``tool_registry.py``.
+moving it would bust the KV cache. The instruction template and its formatter live
+in ``prompt_builder.py`` (``build_rewrite_prompt``) next to the rest of prompt
+assembly; only the Python glue around the tool relocates here — mirroring how
+``length_guard.py`` leaves the ``editor_rewrite`` schema in ``tool_registry.py``.
 """
 
 from __future__ import annotations
@@ -24,22 +25,6 @@ from typing import Any, Iterable, Mapping
 #: The tool name, as a single source of truth for the literal that the schema in
 #: ``tool_registry.py`` registers and that the director/orchestrator key off.
 REWRITE_TOOL_NAME = "rewrite_user_prompt"
-
-
-#: Per-turn request tail handed to the director when running the rewrite tool: the
-#: user's raw message, quoted for the model to refine. Lives here (not in the tools
-#: blob) so it can vary per turn without busting the KV cache — mirroring
-#: ``LENGTH_GUARD_INSTRUCTIONS`` leaving ``tool_registry.py``.
-REWRITE_PROMPT_PROMPT = 'User\'s message:\n"""[{user_message}]"""'
-
-
-def build_rewrite_prompt(user_message: str) -> str:
-    """Format :data:`REWRITE_PROMPT_PROMPT` with the raw *user_message*.
-
-    Appended by :func:`~backend.prompt_builder.build_director_tool_prompt` as the
-    request tail for the ``rewrite_user_prompt`` branch.
-    """
-    return REWRITE_PROMPT_PROMPT.format(user_message=user_message)
 
 
 def extract_rewritten_message(args: Mapping[str, Any]) -> str | None:
