+++ backend/prompt_builder.py
@@ -245,6 +245,21 @@ STRUCTURAL_REWRITE_INSTRUCTIONS = (
     "response is laid out distinctly from the previous ones."
 )
 
+#: Per-turn request tail handed to the director when running the rewrite tool: the
+#: user's raw message, quoted for the model to refine. Lives here (not in the tools
+#: blob) so it can vary per turn without busting the KV cache — mirroring
+#: ``LENGTH_GUARD_INSTRUCTIONS`` leaving ``tool_registry.py``.
+REWRITE_PROMPT_PROMPT = 'User\'s message:\n"""[{user_message}]"""'
+
+
+def build_rewrite_prompt(user_message: str) -> str:
+    """Format :data:`REWRITE_PROMPT_PROMPT` with the raw *user_message*.
+
+    Appended by :func:`build_director_tool_prompt` as the request tail for the
+    ``rewrite_user_prompt`` branch.
+    """
+    return REWRITE_PROMPT_PROMPT.format(user_message=user_message)
+
 
 def build_director_tool_prompt(
     tool_name: str,
@@ -283,12 +298,6 @@ def build_director_tool_prompt(
             parts.append("Previous progressive fields - dynamically update these:\n" + "\n".join(progressive_lines))
         parts.append(f'User\'s next message (for context, take this into account when directing):\n"""{user_message}"""')
     elif tool_name == "rewrite_user_prompt":
-        # Deferred import: the passes depend on prompt_builder (director.py imports
-        # build_director_tool_prompt), so importing prompt_rewrite at module top
-        # would cycle through the passes.director package __init__. Local import
-        # keeps the layering one-directional at load time.
-        from .passes.director.prompt_rewrite import build_rewrite_prompt
-
         parts.append(build_rewrite_prompt(user_message))
     return "\n\n".join(parts)
